Add initial state files
