+++ src/features/documentation/initialState.js
@@ -1,0 +1,12 @@
+const initialState = {
+  impCells: [
+    0,0,0,0,0,
+    0,0,0,0,0,
+    0,0,0,0,0,
+    0,0,0,0,0,
+    0,0,0,0,0
+  ],
+  impIndex: 7
+}
+
+export default initialState
