+++ src/features/feedback/initialState.js
@@ -1,0 +1,5 @@
+const initialState = {
+  feedbackMessage: ''
+}
+
+export default initialState
