+++ src/features/notifications/initialState.js
@@ -1,0 +1,5 @@
+const initialState = {
+  notifications: []
+}
+
+export default initialState
