+++ src/features/parser/reducer.js
@@ -9,21 +9,8 @@ import {
   SET_COLOURS
 } from './actions'
 
-import { defaultWarriors } from '../../helpers/defaultWarriors'
-import blankWarrior from '../../helpers/blankWarrior'
-import { colour } from '../common/theme'
-
 // state
-const initialState = {
-  currentFileIndex: 0,
-  currentWarrior: blankWarrior,
-  warriors: [],
-  warriorLibrary: defaultWarriors,
-  standardId: 2, // TODO: what's the best standard to use as a default?
-  displayConsole: false,
-  displayFileManager: false,
-  colours: colour.warrior
-}
+import initialState from './initialState'
 
 // selectors
 export const getParserState = state => state.parser
