+++ src/features/signup/initialState.js
@@ -1,0 +1,5 @@
+const initialState = {
+  signupMessage: ''
+}
+
+export default initialState
